feat: Add dynamic properties and methods to user object
This commit adds dynamic properties and methods to the `user` object in the `05-objetos/02-dinamico.js` file. The `name` property is set to 'John Doe' and the `guardar` method is defined to log the name of the user. Additionally, the `delete` keyword is used to remove the `name` property and the `guardar` method from the `user` object. The updated `user` object is then logged to the console.